google_auth_login/requirements.txt
Flask==2.0.1
requests==2.25.1
google-cloud-datastore==2.1.6
